requirements.txt
asgiref==3.8.1
boto3==1.37.19
botocore==1.37.19
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
defusedxml==0.7.1
Django==5.1.7
django-cors-headers==4.7.0
django-ses==4.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
s3transfer==0.11.4
setuptools==75.8.0
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
urllib3==2.3.0
wheel==0.45.1
whitenoise==6.9.0
